imageUpscaler/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.8
PyQt5>=5.15.0
tqdm>=4.62.0
requests>=2.25.0
